Decouple "around node" SGs #459; inval edge cts
This seems to be working, but edge orientations seem weird in the
layout for simple-strand-tangled.gfa when you draw around node 2
with d = 1 -- and the code is kind of messy.

I suspect the problem stems from how we special-case inval edges
as a separate thing from Subgraph.edges -- which means that
num_real_edges, total_length, num_positive_names, etc will be
inaccurate. I think the solution will be making DecoupledSubgraph
add invalidated edges as real edges (also add some validation code
that all inval edges are real lol), which will hopefully remove the
need to think about inval edges in DrawResults and MAYYBE to a
lesser extent in Layout.

Some ideas:

- Move inval_edge_info from being a 3-tuple to being a fancy Object
  we can just interact with

- Maybe make an InvalidatedEdge object?

- Make a Subgraph-level method for getting the source and target of
  an edge for drawing. Then, if it is an invalidated edge, you can
  use the inval_edge_info stuff. This is why Layout keeps needing to
  care about has_invals.

  - Also, if we do this, then we might even be able to replace
    DecoupledSubgraph with just another Subgraph? But one thing at
    a time...
