+++ metagenomescope/tests/input/simple-strand-tangled.gfa
@@ -1,0 +1,7 @@
+H	VN:Z:1.0
+S	1	CGATGCAAAG
+S	2	TGCAAAGTAC
+S	3	CGATGCAAAG
+L	3	-	1	+	5M
+L	1	+	2	+	5M
+L	1	+	2	-	5M
